Created Grid layout
Installed tailwind and added additional class for 11 columns.  Created an array of 10 rows starting from 1 and in each row added 10 cells which hold the information of whether there is a ship in there and if it is hit/sunk/miss. Looped over the array to create the html grid and output the letters along the top line and numbers along the side.  The cells will output icons depenidng on their state