randomly added ships to board with function
Removed manual additions of ships to board. Added a new game button, when clicked this runs a function that resets my board and then loops over the ships in the board and runs setShips function which passes down the size and qty of the ships.  This function sets a random number and random letter and then uses that as the first coordinate on the board.  Once the first coordinate is set if a ship is already on that coordinate then another random coordinate will be retrieved.  If there is no ship on the coordinate then a horizontalCheck function is ran, this loops over the next coordinates on in sequence on that row and if they fit on the line i.e. they are not too close to the edge of the board then those coordinates are returned, if they do not fit on the row then a vertical check function is ran and this does the same thing but checks if the cells down the board vertically are available and if so then those coordinates are returned. Those coordinates are then looped over with canPlaceShip and it makes sure that there are no ships on any of those cells and if there are none then the ships is added to each of these cells and the array is updated with that data, if there is a ship on that cell then the startingCell is set to false again the the whole loop is started again form the top setting a new random number and random letter.