Added more information on sidebar and alerts
Added ship type to the ships object, this allows me to output how many ships are on the board in the sidebar and also add the information to the events when a ships has been sank to tell the user what they sank. Added number of attempts user has made and when the user has sank all ships add in a congratulations message along with how many attempts they had and prompt them to beat the score in a new game.