Layout and Game start state
Created a state where when the game starts the user is asked to enter in their Name before they begin and none of the board is shown until they do that then everything is displayed.  The users name is shown on the right.  The buttons have been aligned along with the other content on the page to make it move visually attractive.  When a new game is create I have reset all of the state values